+++ examples/demo.py
@@ -19,7 +19,7 @@ if __name__ == "__main__":
         a 3 4 0 15 4
     """)
 
-    NUM_COMMODITIES = 3
+    NUM_COMMODITIES = 5
     SEED = 512
 
     exp_file = os.path.join(EXAMPLES_DIR, "exp_network.min")
@@ -28,18 +28,25 @@ if __name__ == "__main__":
 
     print(f"Created baseline file: {exp_file}")
 
+    exp_file = os.path.join(EXAMPLES_DIR, "netgen_1.min")
     network = s2mflow.load_min_instance(exp_file)
+    print("network supplies: ")
+    print(network.supplies)
 
     print("-" * 20, "Uniform Partitioning", "-" * 20)
     uniform_mc_data = s2mflow.generate_multi_commodity_data(
         instance=network,
         num_commodities=NUM_COMMODITIES,
-        is_uniform=True,
+        method=1,
         randomize_caps=False,
         randomize_costs=False,
     )
 
-    output_path = os.path.join(EXAMPLES_DIR, "uniform.mcfmin")
+    print("uniform supplies: ")
+    print(uniform_mc_data.supply_partition)
+    print("CDH: ", s2mflow.compute_commodity_demand_heterogeneity(uniform_mc_data.supply_partition, network.supplies))
+
+    output_path = os.path.join(EXAMPLES_DIR, f"netgen_1_uniform_{NUM_COMMODITIES}.mcfmin")
     s2mflow.save_multi_commodity_instance(output_path, network, uniform_mc_data)
 
     loaded_uniform_mc_data = s2mflow.load_multi_commodity_instance(output_path)
@@ -48,22 +55,44 @@ if __name__ == "__main__":
     spread_mc_data = s2mflow.generate_multi_commodity_data(
         instance=network,
         num_commodities=NUM_COMMODITIES,
-        is_uniform=False,
+        method=0,
         randomize_caps=False,
         randomize_costs=False,
         seed=SEED,
     )
+    print("spread supplies: ")
+    print(spread_mc_data.supply_partition)
+    print("CDH: ", s2mflow.compute_commodity_demand_heterogeneity(spread_mc_data.supply_partition, network.supplies))
     
-    output_path = os.path.join(EXAMPLES_DIR, "spread.mcfmin")
+    output_path = os.path.join(EXAMPLES_DIR, f"netgen_1_spread_{NUM_COMMODITIES}.mcfmin")
     s2mflow.save_multi_commodity_instance(output_path, network, spread_mc_data)
 
     loaded_spread_mc_data = s2mflow.load_multi_commodity_instance(output_path)
+
+    print("-" * 20, "Beta-Binomial Partitioning", "-" * 20)
+    beta_binomial_mc_data = s2mflow.generate_multi_commodity_data(
+        instance=network,
+        num_commodities=NUM_COMMODITIES,
+        method=2,
+        randomize_caps=False,
+        randomize_costs=False,
+        concentration_param=30.0,
+        seed=SEED,
+    )
+    print("beta-binomial supplies: ")
+    print(beta_binomial_mc_data.supply_partition)
+    print("CDH: ", s2mflow.compute_commodity_demand_heterogeneity(beta_binomial_mc_data.supply_partition, network.supplies))
+
+    output_path = os.path.join(EXAMPLES_DIR, f"netgen_1_beta_binomial_{NUM_COMMODITIES}.mcfmin")
+    s2mflow.save_multi_commodity_instance(output_path, network, beta_binomial_mc_data)
+
+    loaded_beta_binomial_mc_data = s2mflow.load_multi_commodity_instance(output_path)
 
     print("-" * 20, "Spread Partitioning with randomized capacities", "-" * 20)
     spread_mc_rand_caps_data = s2mflow.generate_multi_commodity_data(
         instance=network,
         num_commodities=NUM_COMMODITIES,
-        is_uniform=False,
+        method=0,
         randomize_caps=True,
         cap_a=0.6,
         cap_b=1.0,
@@ -80,7 +109,7 @@ if __name__ == "__main__":
     spread_mc_rand_costs_data = s2mflow.generate_multi_commodity_data(
         instance=network,
         num_commodities=NUM_COMMODITIES,
-        is_uniform=False,
+        method=0,
         randomize_caps=False,
         randomize_costs=True,
         cost_a=0.5,
@@ -97,7 +126,7 @@ if __name__ == "__main__":
     spread_mc_rand_caps_costs_data = s2mflow.generate_multi_commodity_data(
         instance=network,
         num_commodities=NUM_COMMODITIES,
-        is_uniform=False,
+        method=0,
         randomize_caps=True,
         cap_a=0.6,
         cap_b=1.0,
@@ -112,7 +141,7 @@ if __name__ == "__main__":
 
     loaded_spread_mc_rand_caps_costs_data = s2mflow.load_multi_commodity_instance(output_path)
 
-    data = {1: 126, 2: -126}
+    data = {1: 387, 2: -387}
     print(data)
 
     spread_multi_data = s2mflow.split_supplies_spread(
@@ -126,9 +155,19 @@ if __name__ == "__main__":
         num_commodities=5,
     )
 
+    beta_binomial_multi_data = s2mflow.split_supplies_beta_binomial(
+        data,
+        num_commodities=5,
+        seed=SEED,
+    )
+
     print(spread_multi_data)
+    print(s2mflow.compute_commodity_demand_heterogeneity(spread_multi_data, data))
     print(uniform_multi_data)
+    print(s2mflow.compute_commodity_demand_heterogeneity(uniform_multi_data, data))
+    print(beta_binomial_multi_data)
+    print(s2mflow.compute_commodity_demand_heterogeneity(beta_binomial_multi_data, data))
 
     incoming, outgoing = s2mflow.get_adjacency_mapping(network.nodes, network.arcs)
-    print(incoming)
-    print(outgoing)
+    #print(incoming)
+    #print(outgoing)